backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
pymongo
fastapi-cors